basically redid the way it 'starts'

+++ bucket/bucket.py
@@ -1,16 +1,47 @@
 import os
 import re
-import sentiment
 import subprocess
+import modules.gptj as gptj
 
 def data(text):
     try:
-        label = sentiment.bucket(text)
-        if (("negative" not in label) and ("unwholesome" not in label) and ("mean" not in label) and ("offensive" not in label) and ("horny" not in label)):
-            blacklist = ["cum","fart","sex","serbia","ploopy","greece",
-                "politics","political","capitalist","capitalism","communist",
-                "communism","socialist","socialism","democrat","democracy",
-                "facist","facism","republic","republican"]
+        blacklist = ["cum","fart","sex","serbia","ploopy","greece",
+            "politics","political","capitalist","capitalism","communist",
+            "communism","socialist","socialism","democrat","democracy",
+            "facist","facism","republic","republican"]
+        try:
+            analyze = gptj.Sentiment()
+            labels = (
+                ["positive","neutral","negative"],
+                ["wholesome","neutral","unwholesome"],
+                ["nice","neutral","mean"],
+                ["harmless","neutral","offensive"]
+            )
+            results = analyze.multi(text, labels)
+            classify = []
+            for result in results:
+                tags = []
+                scores = []
+                for labels in result.keys():
+                    tags.append(labels)
+                for probability in result.values():
+                    scores.append(round(probability*100, 2))
+                score = scores[0]
+                tag = tags[0]
+                for n in range(len(tags)):
+                    if scores[n] > score:
+                        score = scores[n]
+                        tag = tags[n]
+                if (score > 60):
+                    classify.append(tag)
+                if (("negative" not in classify) and ("unwholesome" not in classify) and ("mean" not in classify) and ("offensive" not in classify)):
+                    text_ok = True
+                else:
+                    text_ok = False
+        except:
+            text_ok = False
+
+        if (text_ok):
             for word in blacklist:
                 if (word in text):
                     text = re.sub('(?i)'+re.escape(word), lambda m: '[**CENSORED**]', text)
@@ -26,11 +57,15 @@ def data(text):
 
 def reply():
     try:
-        subprocess.getoutput("./bucket/mrkfeed.awk < ./bucket/tempbucket.txt >> ./bucket/model.mrkdb")
-        if os.path.exists("./bucket/tempbucket.txt"):
-            os.remove("./bucket/tempbucket.txt")
-        message = subprocess.getoutput("./bucket/mrkwords.sh ./bucket/model.mrkdb 555|head -c1000|tr '\n' ' ' && echo")
-        return message
+        if os.name == "posix":
+            subprocess.getoutput("chmod +x ./bucket/mrkfeed.awk")
+            subprocess.getoutput("./bucket/mrkfeed.awk < ./bucket/tempbucket.txt >> ./bucket/model.mrkdb")
+            if os.path.exists("./bucket/tempbucket.txt"):
+                os.remove("./bucket/tempbucket.txt")
+            message = subprocess.getoutput("./bucket/mrkwords.sh ./bucket/model.mrkdb 555|head -c1000|tr '\n' ' ' && echo")
+            return message
+        else:
+            return
     except BaseException as error:
         print("\n----ERROR----\nfailed 'BUCKET REPLY'\n"+str(error))
         return
